only set upid.Host and upid.Port if Host is unset
This commit checks if `upid.Host` is set in the http_transporter and sets it
to the value that it determines if unset.

The original problem was that I was trying to run a scheduler on 0.0.0.0 as
listening host and have OverrideHostname set, this should have led to the
master thinking that I was sending messages from HostnameOverride, instead the
http_transporter function listen(), mutated the upid.Host and the master would
always look up 0.0.0.0.

The patch has been tested by running a scheduler against a Mesos cluster with
the scheduler bound to 0.0.0.0 and HostnameOverride in DriverConfig to an
external IP.