messenger: Defer free port allocation to kernel
This commit fixes removes some non-determinism in the messenger package tests
which caused them to fail with port allocation conflicts. The removed
test case in `TestMutatedHostUPid` is required since we can't know the
allocated port in advance.

We're removing `TestMessengerFailToStart` because testing that a kernel doesn't
allow the two sockets to bind to a single port (unless using SO_REUSEPORT in Linux)
isn't the responsibility of a user-space program.

Fixes #169