Modified Makefile to test only directories with tests available
At the moment, tests are run on *every* subpackage within mesos-go,
which generates a lot of failures since tests aren't available in many
packages at the moment. This diff changes the test target to only run
tests on subpackages with tests available so we get more signal from the
tests.