controller: WithContextPerSubscription()
Prior to this changeset, it is difficult for event handlers to know when
the current subscription has terminated. This is important, especially,
for offer processing since all outstanding offers on a terminated
subscription are immediately invalidated by Mesos. This changeset allows
a scheduler to bound the cancelation context to the lifetime of the
current subscription.